requirements.txt
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
filelock==3.0.12
idna==2.10
joblib==0.17.0
nltk==3.5
pandas==1.1.3
pipenv==2020.8.13
protobuf==3.17.2
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.10.15
requests==2.24.0
scikit-learn==0.22
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.50.2
urllib3==1.25.10
vaderSentiment==3.3.2
virtualenv==20.0.35
virtualenv-clone==0.5.4
zipp==3.3.0
